Improve restoration of selections after staging
The ‘try exec’ in the mappings of a and r silence the failure that
happens when the selections cannot be restored.

Running ‘execute-keys %’ and ‘kit-select’ before trying to
restore selections is necessary so that matches outside of the status
area are not selected too (for example if a pathspec occurs in a recent
commit message).